web sitesi/kapadokya/python_api/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
torch
torchvision
Pillow
python-multipart
